Update to 0.2.
Changes:

Fixes #2 by using `git -C PATH` instead of changing the working directory every
time.
Thanks to @mountainash for reporting and providing a first patch.

Update development dependencies that make GitHub's dependabot happy.
Notice that this plugin itself only depends on lektor directly, so any reported
vulnerabilities in dependencies are to be solved in how you install lektor.

This version also fixes a strange issue on BSD that would make the plugin
unusable.
In order to solve that, we now rely directly on `subprocess.run` instead of
lektor's Command class.