Release 0.3 with a bugfix.
Under certain conditions lektor instantiates the plugin multiple times and the
code that checks for the project's root_path resulted in an exception.

After inspecting lektor's source code, we realised that a reference to the
plugin object in the publisher was not necessary (and that the answer to the
question "is this safe?" was a soudn: no).

This was an issue in certain setups, for more information see:
https://github.com/CamilionEU/lektor-git-src-publisher/issues/2

Thanks to @mountainash for reporting, and for their patience in testing and
reaching a working fix for all known use-cases.